requirements.txt
bidict==0.21.2
click==7.1.2
h11==0.12.0
netifaces==0.10.6
python-engineio==4.0.0
python-socketio==5.0.4
starlette==0.14.2
typing-extensions==3.7.4.3
uvicorn==0.13.3
